+++ themes/theme-unicorn/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @marigold/theme-unicorn
 
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`51132dd8`](https://github.com/marigold-ui/marigold/commit/51132dd83ead1b5b5a89b1281ba1a2c2d4e17f04)]:
+  - @marigold/components@1.2.2
+
 ## 6.0.1
 
 ### Patch Changes